requirements.txt
discord.py-self
yt-dlp
ffmpeg-python